# services/device-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Manages sensors, gateways, sensor profiles and sensor groups.",
        "title": "Device Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/api/v1/device",
    "paths": {
        "/company-config": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-config"
                ],
                "summary": "Create company config",
                "parameters": [
                    {
                        "description": "Company config payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCompanyConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCompanyConfigResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/company-config/{companyID}": {
            "delete": {
                "tags": [
                    "company-config"
                ],
                "summary": "Delete company config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company ID",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/gateways": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gateways"
                ],
                "summary": "List all gateways",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GatewayListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "gateways"
                ],
                "summary": "Create a gateway",
                "parameters": [
                    {
                        "description": "Gateway payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateGatewayRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/gateways/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "gateways"
                ],
                "summary": "Update a gateway",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gateway ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateGatewayRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "gateways"
                ],
                "summary": "Delete a gateway",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gateway ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/sensor-profiles": {
            "get": {
                "description": "Not authenticated — every user can access this list.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sensor-profiles"
                ],
                "summary": "List all sensor profiles",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max number of profiles to return",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SensorProfileListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/sensors": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sensors"
                ],
                "summary": "Lists all sensors.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SensorListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "sensors"
                ],
                "summary": "Create a sensor",
                "parameters": [
                    {
                        "description": "Sensor payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSensorRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/sensors/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "sensors"
                ],
                "summary": "Update a sensor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sensor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateSensorRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "sensors"
                ],
                "summary": "Delete a sensor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SensorID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateCompanyConfigRequest": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.CreateCompanyConfigResponse": {
            "type": "object",
            "properties": {
                "tenant_id": {
                    "type": "string"
                }
            }
        },
        "dto.CreateGatewayRequest": {
            "type": "object",
            "required": [
                "company_id",
                "gateway_eui",
                "name"
            ],
            "properties": {
                "company_id": {
                    "type": "string"
                },
                "gateway_eui": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.CreateSensorRequest": {
            "type": "object",
            "required": [
                "app_key",
                "company_id",
                "device_eui",
                "device_profile_id",
                "factory_id",
                "name"
            ],
            "properties": {
                "app_key": {
                    "type": "string"
                },
                "company_id": {
                    "description": "TODO: CompanyID should be extracted from auth",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "device_eui": {
                    "type": "string"
                },
                "device_profile_id": {
                    "type": "string"
                },
                "factory_area_id": {
                    "description": "Optional — UUID, omit if service not yet available",
                    "type": "string"
                },
                "factory_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "production_resource": {
                    "description": "Optional — UUID, omit if service not yet available",
                    "type": "string"
                }
            }
        },
        "dto.GatewayListResponse": {
            "type": "object",
            "properties": {
                "gateways": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GatewayResponse"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "dto.GatewayResponse": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "factory_area_id": {
                    "type": "string"
                },
                "gateway_eui": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_seen_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.SensorListResponse": {
            "type": "object",
            "properties": {
                "sensors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SensorResponse"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "dto.SensorProfileListResponse": {
            "type": "object",
            "properties": {
                "sensor_profiles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SensorProfileResponse"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "dto.SensorProfileResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "mac_version": {
                    "description": "LoRaWAN version",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "region": {
                    "description": "LoRaWAN region (EU868)",
                    "type": "string"
                },
                "vendor_id": {
                    "description": "Identification of model producer",
                    "type": "string"
                },
                "vendor_name": {
                    "description": "Vendor name",
                    "type": "string"
                }
            }
        },
        "dto.SensorResponse": {
            "type": "object",
            "properties": {
                "app_key": {
                    "type": "string"
                },
                "company_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "device_eui": {
                    "type": "string"
                },
                "device_profile_id": {
                    "type": "string"
                },
                "factory_area_id": {
                    "type": "string"
                },
                "factory_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_seen_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "production_resource": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateSensorRequest": {
            "type": "object",
            "required": [
                "device_profile_id",
                "factory_id",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "device_profile_id": {
                    "type": "string"
                },
                "factory_area_id": {
                    "description": "Optional — UUID, omit if service not yet available",
                    "type": "string"
                },
                "factory_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "production_resource": {
                    "description": "Optional — UUID, omit if service not yet available",
                    "type": "string"
                }
            }
        }
    }
}